REF: misplaced CategoricalIndex tests
