+++ pandas/tests/indexing/test_categorical.py
@@ -435,32 +435,6 @@ class TestCategoricalIndex:
         with pytest.raises(KeyError, match=msg):
             df.loc[["a", "x"]]
 
-    def test_get_indexer_array(self):
-        arr = np.array(
-            [Timestamp("1999-12-31 00:00:00"), Timestamp("2000-12-31 00:00:00")],
-            dtype=object,
-        )
-        cats = [Timestamp("1999-12-31 00:00:00"), Timestamp("2000-12-31 00:00:00")]
-        ci = CategoricalIndex(cats, categories=cats, ordered=False, dtype="category")
-        result = ci.get_indexer(arr)
-        expected = np.array([0, 1], dtype="intp")
-        tm.assert_numpy_array_equal(result, expected)
-
-    def test_get_indexer_same_categories_same_order(self):
-        ci = CategoricalIndex(["a", "b"], categories=["a", "b"])
-
-        result = ci.get_indexer(CategoricalIndex(["b", "b"], categories=["a", "b"]))
-        expected = np.array([1, 1], dtype="intp")
-        tm.assert_numpy_array_equal(result, expected)
-
-    def test_get_indexer_same_categories_different_order(self):
-        # https://github.com/pandas-dev/pandas/issues/19551
-        ci = CategoricalIndex(["a", "b"], categories=["a", "b"])
-
-        result = ci.get_indexer(CategoricalIndex(["b", "b"], categories=["b", "a"]))
-        expected = np.array([1, 1], dtype="intp")
-        tm.assert_numpy_array_equal(result, expected)
-
     def test_getitem_with_listlike(self):
         # GH 16115
         cats = Categorical([Timestamp("12-31-1999"), Timestamp("12-31-2000")])
